# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zhalwadi Samaj API Documentation",
        "description": "API documentation for Zhalwadi Samaj mobile application",
        "contact": {
            "email": "support@zhalwadisamaj.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://parivar.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/member_gallery/get_all": {
            "get": {
                "tags": [
                    "Member Gallery"
                ],
                "summary": "Get All Member Galleries",
                "operationId": "e65c5d040e71ad69d7c1f444e9cb30c6",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nativePlaces/get_all": {
            "get": {
                "tags": [
                    "Native Places"
                ],
                "summary": "Get All Native Places",
                "operationId": "8011f427e4f5e5ff321664e295319b37",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/result/get_all": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "Get All Results",
                "operationId": "22fc2481fe20f28af123f58de90efc80",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/team/get_all": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get All Teams",
                "operationId": "c823cf3fdd2684ab1e868c13064fbbf2",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Zone/get_all": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Get All Zones",
                "operationId": "7d7d262603902c2fe59e83607c68865c",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your bearer token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}